Добавил изменения bot.models.py поле user_id bigField

+++ bot/models.py
@@ -10,7 +10,7 @@ class UsersBot(models.Model):
     datetime дата
     ls список лицевых ['123445678','87654321']
     """
-    user_id = models.IntegerField(verbose_name='ID Telegram')
+    user_id = models.BigIntegerField(verbose_name='ID Telegram')
     username = models.CharField(max_length=100, blank=True, default=None, verbose_name='Имя пользователя telegram')
     datetime = models.DateTimeField(auto_now_add=True, verbose_name='Дата добавления лицевого')
     ls = models.IntegerField(verbose_name='Лицевой счет')
